refactor: move source interface to types package
