+++ pkg/types/types.go
@@ -1,7 +1,7 @@
 // Copyright 2022 D2iQ, Inc. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package sources
+package types
 
 // Source provides an interface for managing a plugin source.
 type Source interface {
